Refactoring for pay roll run

+++ CorpMediaJson/src/com/intuiture/corp/json/DailyAllowanceJson.java
@@ -3,7 +3,7 @@ package com.intuiture.corp.json;
 public class DailyAllowanceJson {
 
 	private Integer dailyAllowanceId;
-	private Double maxAnnualLimit;
+	private Long maxAnnualLimit;
 	private Integer requireSubmissionId;
 	private Integer companyId;
 
@@ -15,11 +15,11 @@ public class DailyAllowanceJson {
 		this.dailyAllowanceId = dailyAllowanceId;
 	}
 
-	public Double getMaxAnnualLimit() {
+	public Long getMaxAnnualLimit() {
 		return maxAnnualLimit;
 	}
 
-	public void setMaxAnnualLimit(Double maxAnnualLimit) {
+	public void setMaxAnnualLimit(Long maxAnnualLimit) {
 		this.maxAnnualLimit = maxAnnualLimit;
 	}
